requirements.txt
aiogram==3.13.1
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
duckdb>=0.9.0
fastapi>=0.104.0
twilio>=8.0.0
pytest>=7.0.0
pathlib 